Rename `/me` to `/settings`
Only `me.index` is migrated to `settings.index`, the other nested `me` routes still keep their old paths for now.
