+++ app/controllers/settings/index.js
@@ -6,7 +6,7 @@ import { alias } from 'macro-decorators';
 
 import ajax from '../../utils/ajax';
 
-export default class MeIndexController extends Controller {
+export default class SettingsController extends Controller {
   isResetting = false;
 
   @alias('model.ownedCrates') ownedCrates;
