+++ app/routes/me/index.js
@@ -1,29 +1,10 @@
+import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
 
-import AuthenticatedRoute from '../-authenticated-route';
+export default class MeIndexRoute extends Route {
+  @service router;
 
-export default class MeIndexRoute extends AuthenticatedRoute {
-  @service store;
-
-  async model() {
-    let { ownedCrates, currentUser: user } = this.session;
-
-    if (!ownedCrates) {
-      await this.session.fetchUser();
-      ({ ownedCrates } = this.session);
-    }
-
-    let apiTokens = await this.store.findAll('api-token');
-
-    return { user, ownedCrates, api_tokens: apiTokens.toArray() };
-  }
-
-  setupController(controller) {
-    super.setupController(...arguments);
-
-    controller.setProperties({
-      emailNotificationsSuccess: false,
-      emailNotificationsError: false,
-    });
+  redirect() {
+    this.router.replaceWith('settings');
   }
 }
